refactor(systems): refined cast safety and parallelization metrics
- Fix potential casting overflows in AVX2 distance kernels
- Update documentation with more accurate parallelization metrics (~3.3x)
- Add try_from for safe index-to-ID conversion in KMeansRouter

+++ src/graph/mod.rs
@@ -148,10 +148,11 @@ impl HNSWGraph {
                 let vxor = _mm256_xor_si256(va, vb);
                 // On x86, we don't have a 256-bit popcount, so we use 64-bit ones
                 // or specialized libraries. Here we stick to a middle ground.
-                let x0 = _mm256_extract_epi64(vxor, 0) as u64;
-                let x1 = _mm256_extract_epi64(vxor, 1) as u64;
-                let x2 = _mm256_extract_epi64(vxor, 2) as u64;
-                let x3 = _mm256_extract_epi64(vxor, 3) as u64;
+                // Extract as i64 and reinterpret bits as u64 (no sign loss, just bit pattern)
+                let x0 = _mm256_extract_epi64(vxor, 0) as i64 as u64;
+                let x1 = _mm256_extract_epi64(vxor, 1) as i64 as u64;
+                let x2 = _mm256_extract_epi64(vxor, 2) as i64 as u64;
+                let x3 = _mm256_extract_epi64(vxor, 3) as i64 as u64;
                 total += x0.count_ones() + x1.count_ones() + x2.count_ones() + x3.count_ones();
             }
             i += 4;
